Implemented options for colour selection. Options do not affect actual plot yet
